+++ app/Mail/UserDelete.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\Mail;
+
+use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Mail\Mailable;
+use Illuminate\Queue\SerializesModels;
+
+class UserRegister extends Mailable
+{
+    use Queueable, SerializesModels;
+
+    public $token;
+    public $email;
+
+    public function __construct($token, $email)
+    {
+        $this->token = $token;
+        $this->email = $email;
+    }
+
+    public function build()
+    {
+        return $this
+            ->subject('Account Deletion Confirmation')
+            ->markdown('emails.user_delete')
+            ->with([
+                'email' => $this->email,
+            ]);
+    }
+}